add cow shed related abi and move contracts codegen folder to common
