+++ abi/CoWShed.json
@@ -1,0 +1,139 @@
+[
+  { "inputs": [], "name": "AlreadyInitialized", "type": "error" },
+  { "inputs": [], "name": "DeadlineElapsed", "type": "error" },
+  { "inputs": [], "name": "InvalidSignature", "type": "error" },
+  { "inputs": [], "name": "InvalidSignature", "type": "error" },
+  { "inputs": [], "name": "NonceAlreadyUsed", "type": "error" },
+  { "inputs": [], "name": "OnlyAdmin", "type": "error" },
+  { "inputs": [], "name": "OnlyAdminOrTrustedExecutorOrSelf", "type": "error" },
+  { "inputs": [], "name": "OnlySelf", "type": "error" },
+  { "inputs": [], "name": "OnlyTrustedExecutor", "type": "error" },
+  {
+    "anonymous": false,
+    "inputs": [
+      { "indexed": false, "internalType": "bytes32", "name": "r", "type": "bytes32" },
+      { "indexed": false, "internalType": "bytes32", "name": "s", "type": "bytes32" },
+      { "indexed": false, "internalType": "uint8", "name": "v", "type": "uint8" }
+    ],
+    "name": "RSV",
+    "type": "event"
+  },
+  {
+    "anonymous": false,
+    "inputs": [
+      { "indexed": false, "internalType": "address", "name": "previousExecutor", "type": "address" },
+      { "indexed": false, "internalType": "address", "name": "newExecutor", "type": "address" }
+    ],
+    "name": "TrustedExecutorChanged",
+    "type": "event"
+  },
+  {
+    "anonymous": false,
+    "inputs": [{ "indexed": true, "internalType": "address", "name": "implementation", "type": "address" }],
+    "name": "Upgraded",
+    "type": "event"
+  },
+  {
+    "inputs": [{ "internalType": "address", "name": "resolver", "type": "address" }],
+    "name": "claimWithResolver",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [],
+    "name": "domainSeparator",
+    "outputs": [{ "internalType": "bytes32", "name": "", "type": "bytes32" }],
+    "stateMutability": "view",
+    "type": "function"
+  },
+  {
+    "inputs": [
+      {
+        "components": [
+          { "internalType": "address", "name": "target", "type": "address" },
+          { "internalType": "uint256", "name": "value", "type": "uint256" },
+          { "internalType": "bytes", "name": "callData", "type": "bytes" },
+          { "internalType": "bool", "name": "allowFailure", "type": "bool" },
+          { "internalType": "bool", "name": "isDelegateCall", "type": "bool" }
+        ],
+        "internalType": "struct Call[]",
+        "name": "calls",
+        "type": "tuple[]"
+      },
+      { "internalType": "bytes32", "name": "nonce", "type": "bytes32" },
+      { "internalType": "uint256", "name": "deadline", "type": "uint256" },
+      { "internalType": "bytes", "name": "signature", "type": "bytes" }
+    ],
+    "name": "executeHooks",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [
+      { "internalType": "address", "name": "factory", "type": "address" },
+      { "internalType": "bool", "name": "claimResolver", "type": "bool" }
+    ],
+    "name": "initialize",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [{ "internalType": "bytes32", "name": "nonce", "type": "bytes32" }],
+    "name": "nonces",
+    "outputs": [{ "internalType": "bool", "name": "", "type": "bool" }],
+    "stateMutability": "view",
+    "type": "function"
+  },
+  {
+    "inputs": [{ "internalType": "bytes32", "name": "nonce", "type": "bytes32" }],
+    "name": "revokeNonce",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [
+      {
+        "components": [
+          { "internalType": "address", "name": "target", "type": "address" },
+          { "internalType": "uint256", "name": "value", "type": "uint256" },
+          { "internalType": "bytes", "name": "callData", "type": "bytes" },
+          { "internalType": "bool", "name": "allowFailure", "type": "bool" },
+          { "internalType": "bool", "name": "isDelegateCall", "type": "bool" }
+        ],
+        "internalType": "struct Call[]",
+        "name": "calls",
+        "type": "tuple[]"
+      }
+    ],
+    "name": "trustedExecuteHooks",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [],
+    "name": "trustedExecutor",
+    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
+    "stateMutability": "view",
+    "type": "function"
+  },
+  {
+    "inputs": [{ "internalType": "address", "name": "newImplementation", "type": "address" }],
+    "name": "updateImplementation",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [{ "internalType": "address", "name": "who", "type": "address" }],
+    "name": "updateTrustedExecutor",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  { "stateMutability": "payable", "type": "receive" }
+]
