+++ src/composable/utils.ts
@@ -4,10 +4,10 @@ import {
   EXTENSIBLE_FALLBACK_HANDLER_CONTRACT_ADDRESS,
   SupportedChainId,
 } from '../common'
-import { ExtensibleFallbackHandler__factory } from './generated'
+import { ExtensibleFallbackHandler__factory } from '../common/generated'
 import { BlockInfo, ConditionalOrderParams } from './types'
 import { Order, OrderBalance, OrderKind } from '@cowprotocol/contracts'
-import { GPv2Order } from './generated/ComposableCoW'
+import { GPv2Order } from '../common/generated/ComposableCoW'
 
 const ERC20_BALANCE_VALUES = ['erc20', '0x5a28e9363bb942b639270062aa6bb295f434bcdfc42c97267bf003f272060dc9']
 const EXTERNAL_BALANCE_VALUES = ['external', '0xabee3b73373acd583a130924aad6dc38cfdc44ba0555ba94ce2ff63980ea0632']
